fix u.Bool again

+++ Cast.go
@@ -190,7 +190,7 @@ func String(value interface{}) string {
 func Bool(value interface{}) bool {
 	switch realValue := value.(type) {
 	case int, int8, int16, int32, int64, uint, uint8, uint16, uint32, uint64, float32, float64:
-		return realValue != 0
+		return Uint64(realValue) != 0
 	case bool:
 		return realValue
 	case []byte:
